% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_cell_empty}
\alias{is_cell_empty}
\title{Check if a dataframe cell is empty}
\usage{
is_cell_empty(cell)
}
\arguments{
\item{cell}{A single value from a dataframe.}
}
\value{
`TRUE` if the cell is empty (`NA` or `""`), otherwise `FALSE`.
}
\description{
This function checks whether a given cell from a dataframe is empty.
A cell is considered empty if it contains `NA` or an empty string (`""`).
}
\keyword{internal}
