% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{validate_fk}
\alias{validate_fk}
\title{Validate foreign key column}
\usage{
validate_fk(table_pk, name_pk, table_fk, name_fk)
}
\arguments{
\item{table_pk}{A data frame that represents the table containing the primary key.}

\item{name_pk}{A string indicating the name of the table to which reference.}

\item{table_fk}{A data frame that represents the table containing the foreign key.}

\item{name_fk}{A character string indicating the name of the reference table.}
}
\value{
`TRUE`. Generates warnings if the foreign key column is missing from the
reference table or if it contains values that are not present in the column primary
key of the referenced table.
}
\description{
Checks if a foreign key column exists in the reference table and ensures make all
values in the foreign key column match the existing values in the primary key
column of the referenced table.
}
\keyword{internal}
