% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_tab}
\alias{check_tab}
\title{Comprueba que todos los tokens estén definidos en las tablas.
Comprueba la relación entre claves primarias y foráneas.
Avisa de todos los errores encontrados.}
\usage{
check_tab(file_path, template_path)
}
\arguments{
\item{file_path}{Character. Path to the spreadsheet file to be converted.}

\item{template_path}{Character. Path to the XML template file to guide the
transformation process.}
}
\description{
Comprueba que todos los tokens estén definidos en las tablas.
Comprueba la relación entre claves primarias y foráneas.
Avisa de todos los errores encontrados.
}
